+++ source/redux/client/client.js
@@ -1,20 +1,20 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
 
-import Http_client from '../http client'
-import localize_and_render from '../client'
+import Http_client from '../../http client'
+import localize_and_render from '../../client'
 
-import { render_on_client } from './render'
+import render_on_client from './render'
 import create_store from './store'
-import { normalize_common_options } from './normalize'
-import set_up_http_client from './http client'
+import { normalize_common_options } from '../normalize'
+import set_up_http_client from '../http client'
 
 // Performs client-side rendering
 // along with varios stuff like loading localized messages.
 //
 // This function is what's gonna be called from the project's code on the client-side.
 //
-export default function render({ development, development_tools, load_translation }, common)
+export default function render({ development, development_tools, translation }, common)
 {
 	common = normalize_common_options(common)
 
@@ -32,7 +32,7 @@ export default function render({ development, development_tools, load_translatio
 		middleware           : common.redux_middleware,
 		on_store_created     : common.on_store_created,
 		promise_event_naming : common.promise_event_naming,
-		on_preload_error     : common.preload && common.preload.on_error,
+		on_preload_error     : common.preload && common.preload.catch,
 		preload_helpers      : common.preload && common.preload.helpers,
 		create_routes        : common.create_routes,
 		on_navigate          : common.on_navigate,
@@ -54,7 +54,7 @@ export default function render({ development, development_tools, load_translatio
 	return localize_and_render
 	({
 		development,
-		load_translation,
+		translation,
 		wrapper: common.wrapper,
 		render_on_client,
 		render_parameters: { development_tools, create_routes: common.create_routes, store }